% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft-history.R
\name{draft_history}
\alias{draft_history}
\title{Fantasy draft history}
\usage{
draft_history(lid, old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID.}

\item{old}{If \code{FALSE} (default), return only the current season's draft
history. If \code{TRUE}, return a list with all prior drafts.}

\item{...}{Additional API query arguments passed to \code{\link[httr:GET]{httr::GET()}}.}
}
\value{
A tibble (or list) of draft picks.
}
\description{
Return a table of draft picks by season. For snake drafts, display the round
and pick of each player; for auctions (salary cap), the nominating team and
winning bid.
}
\examples{
draft_history(lid = 252353)
}
