% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-lid.R
\name{extract_lid}
\alias{extract_lid}
\title{Extract ESPN league ID.}
\usage{
extract_lid(url)
}
\arguments{
\item{url}{Any ESPN leuage page URL.}
}
\value{
An ESPN league ID.
}
\description{
From any ESPN league page URL (e.g.,
https://fantasy.espn.com/football/team?leagueId=252353&teamId=6&seasonId=2019),
extract the league ID digits.
}
\examples{
extract_lid("https://fantasy.espn.com/football/team?leagueId=252353&teamId=6&seasonId=2019")
}
