% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy-data.R
\name{fantasy_data}
\alias{fantasy_data}
\title{Get Fantasy Data}
\usage{
fantasy_data(lid, ..., current = TRUE)
}
\arguments{
\item{lid}{The ESPN Leaugue ID.}

\item{...}{Query arguments passed to \code{\link[httr:GET]{httr::GET()}} \code{query} argument (coerced
as a single list).}

\item{current}{logical; if \code{TRUE} the v3 API (for 2019 and later) will be
used; if \code{FALSE} the \emph{historical} API will be used.}
}
\value{
A nested list of API results.
}
\description{
Call \code{\link[httr:GET]{httr::GET()}} on the ESPN Fantasy Football API.
}
\examples{
fantasy_data(lid = 252353, view = "members")
}
