% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{fee_settings}
\alias{fee_settings}
\title{League finance settings}
\usage{
fee_settings(leagueId = ffl_id(), leagueHistory = FALSE, ...)
}
\arguments{
\item{leagueId}{Numeric league ID or ESPN fantasy page URL. Defaults to
\code{getOption("fflr.leagueId")}. Function fails if no ID is found.}

\item{leagueHistory}{logical; Should the \code{leagueHistory} version of the API
be called? If \code{TRUE}, a list of results is returned, with one element for
each historical year of the league.}

\item{...}{Arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}} for parsing.}
}
\value{
A data frame of finance settings by season.
}
\description{
The off-site fees assigned to various roster movies and transactions.
}
\examples{
fee_settings(leagueId = "42654852")
fee_settings(leagueId = "252353", leagueHistory = TRUE)
}
\seealso{
Other League settings: 
\code{\link{draft_settings}()},
\code{\link{league_info}()},
\code{\link{league_name}()},
\code{\link{league_size}()},
\code{\link{roster_settings}()},
\code{\link{schedule_settings}()},
\code{\link{score_settings}()},
\code{\link{trade_settings}()},
\code{\link{waiver_settings}()}
}
\concept{League settings}
