% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffl_get.R
\name{ffl_get}
\alias{ffl_get}
\title{Get Fantasy Current Data}
\usage{
ffl_get(lid, ...)
}
\arguments{
\item{lid}{The ESPN Leaugue ID.}

\item{...}{Query arguments passed to \code{\link[httr:GET]{httr::GET()}} \code{query} argument (coerced
as a single list).}
}
\value{
A nested list of API results.
}
\description{
Call \code{\link[httr:GET]{httr::GET()}} on the ESPN Fantasy Football API.
}
\examples{
ffl_get(lid = 252353, view = "members")
}
