% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffl-merge.R
\name{ffl_merge}
\alias{ffl_merge}
\title{Merge fantasy data}
\usage{
ffl_merge(x, y, ...)
}
\arguments{
\item{x, y}{Fantasy data frames to be merged together.}

\item{...}{Arguments passed to \code{\link[=merge]{merge()}}.}
}
\value{
A merged tibble.
}
\description{
This is a shortcut for the base \code{\link[=merge]{merge()}} method that does not reorder
columns or rows and returns the merged data as a tibble.
}
\examples{
team_roster(lid = 252353)
}
