% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffl-info.R
\name{ffl_season}
\alias{ffl_season}
\title{Current fantasy season}
\usage{
ffl_season(offset = 0)
}
\arguments{
\item{offset}{Add negative or positive values.}
}
\description{
Current fantasy season
}
\examples{
Sys.time()
ffl_season()
}
