% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{ffl_year}
\alias{ffl_year}
\title{Current fantasy season}
\usage{
ffl_year(offset = 0)
}
\arguments{
\item{offset}{Add negative or positive values.}
}
\value{
The current or offset \code{seasonId} integer.
}
\description{
Current fantasy season
}
\examples{
Sys.time()
ffl_year()
}
