% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fflr_api}
\alias{fflr_api}
\title{Call the ESPN Fantasy API}
\usage{
fflr_api(path = "", query = NULL, leagueId = fflr_id())
}
\arguments{
\item{path}{Character vector of API path elements to append to
\verb{/apis/v3/games/ffl}.}

\item{query}{Additional queries also passed, possibly \code{seasonId} or similar.}

\item{leagueId}{Your 6-digit league ID. Defaults to \code{Sys.getenv("ESPN_ID")};
use \code{\link[=Sys.setenv]{Sys.setenv()}} or edit your \code{.Renviron} file to set a default ID.
Alternatively, pass any URL from an ESPN fantasy website and automatically
extract the ID.}

\item{...}{Arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}} for parsing.}
}
\value{
A single JSON string.
}
\description{
Use \code{\link[httr:GET]{httr::GET()}} to make a request to the ESPN API and return the parsed
list from the JSON data. The function uses \code{\link[httr:RETRY]{httr::RETRY()}}, so the call will
repeat up to three times if there is a failure.
}
\examples{
\dontrun{
fflr_api()
}
}
\keyword{internal}
