% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{fflr_id}
\alias{fflr_id}
\title{Find your ESPN fantasy league ID}
\usage{
fflr_id(leagueId = Sys.getenv("ESPN_ID"))
}
\arguments{
\item{leagueId}{Your 6-digit league ID. Defaults to \code{Sys.getenv("ESPN_ID")};
use \code{\link[=Sys.setenv]{Sys.setenv()}} or edit your \code{.Renviron} file to set a default ID.
Alternatively, pass any URL from an ESPN fantasy website and automatically
extract the ID.}
}
\description{
Find your ESPN fantasy league ID
}
