% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_matchup.R
\name{form_matchup}
\alias{form_matchup}
\title{Format a Weekly Matchup}
\usage{
form_matchup(data)
}
\arguments{
\item{data}{The list object returned by \code{\link[=fantasy_roster]{fantasy_roster()}}.}
}
\value{
A two-row tibble fora matchup's \code{HOME} and \code{AWAY} scores.
}
\description{
Take a weekly matchup and return tidy tibble of scores.
}
\examples{
data <- ffl_get(lid = 252353, view = "mMatchup")
matchup <- data$schedule[[5]]
weekly_matchup(matchup)
}
