% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_roster.R
\name{form_roster}
\alias{form_roster}
\title{Format a Team Roster}
\usage{
form_roster(team)
}
\arguments{
\item{team}{The list object returned by \code{\link[=fantasy_roster]{fantasy_roster()}}.}
}
\value{
A tibble with a row for every roster entry.
}
\description{
Take a team's list roster entries and applies \code{\link[=roster_entry]{roster_entry()}}
to return a single tibble per team.
}
\examples{
data <- fantasy_roster(lid = 252353, scoringPeriodId = 3)
team <- data$teams[[5]]
form_roster(team)
}
