% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form-teams.R
\name{form_teams}
\alias{form_teams}
\title{Create ESPN Teams Tibble}
\usage{
form_teams(data)
}
\arguments{
\item{data}{The list object returned by \code{\link[=fantasy_members]{fantasy_members()}}.}
}
\value{
A tibble of teams and owners.
}
\description{
From an ESPN "members" API (via \code{\link[=fantasy_members]{fantasy_members()}}), create a
tidy tibble of league members and their fantasy teams.
}
\examples{
form_teams(data = fantasy_members(252353))
}
