% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tibble.R
\name{form_tibble}
\alias{form_tibble}
\title{Turn Nested Lists to Tibble Rows}
\usage{
form_tibble(l)
}
\arguments{
\item{l}{A list object whose sub-elements should be converted to rows and
bound into a single tibble.}
}
\value{
A nested list of API results.
}
\description{
Call \code{\link[purrr:map_df]{purrr::map_df()}} and \code{\link[tibble:as_tibble]{tibble::as_tibble()}} on a nested
list.
}
\examples{
m <- ffl_get(lid = 252353)
form_tibble(l = m$members)
}
