% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaa_teams}
\alias{gaa_teams}
\title{2015-2020 GAA Teams}
\format{
A data frame with 13 rows and 3 variables:
\describe{
\item{team}{Unique yearly team ID}
\item{abbrev}{Team abbreviation}
\item{years}{Nested list of active years}
...
}
}
\source{
\url{https://fantasy.espn.com/apis/v3/games/ffl/seasons/2020/segments/0/leagues/252353}
}
\usage{
gaa_teams
}
\description{
The GAA is the fantasy league of this package's author.
}
\keyword{datasets}
