% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-teams.R
\name{id2abbrev}
\alias{id2abbrev}
\title{Convert team ID to abbreviation}
\usage{
id2abbrev(lid = NULL, id, ...)
}
\arguments{
\item{lid}{ESPN League ID passed to \code{\link[=league_teams]{league_teams()}} or the table itself.}

\item{id}{A integer vector of team IDs to convert.}

\item{...}{Arguments passed to \code{\link[=league_teams]{league_teams()}}.}
}
\value{
A character vector of team abbreviations.
}
\description{
Convert team ID to abbreviation
}
\examples{
id2abbrev(252353, id = 5)
}
