% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\name{league_standings}
\alias{league_standings}
\title{League standings}
\usage{
league_standings(leagueId = ffl_id(), leagueHistory = FALSE, ...)
}
\arguments{
\item{leagueId}{Numeric league ID or ESPN fantasy page URL. Defaults to
\code{getOption("fflr.leagueId")}. Function fails if no ID is found.}

\item{leagueHistory}{logical; Should the \code{leagueHistory} version of the API
be called? If \code{TRUE}, a list of results is returned, with one element for
each historical year of the league.}

\item{...}{Additional queries passed to \code{\link[httr:GET]{httr::GET()}}. Arguments are
converted to a named list and passed to \code{query} alongside \code{view}.}
}
\value{
A data frame of team standings.
}
\description{
Return the current and projected standings, win streak, total wins, losses,
and points scored for and against each team.
}
\examples{
league_standings(leagueId = "42654852")
}
\seealso{
Other league functions:
\code{\link[=draft_recap]{draft_recap()}},
\code{\link[=league_members]{league_members()}},
\code{\link[=league_messages]{league_messages()}},
\code{\link[=league_status]{league_status()}},
\code{\link[=league_teams]{league_teams()}},
\code{\link[=tidy_schedule]{tidy_schedule()}},
\code{\link[=transaction_counter]{transaction_counter()}}
}
\concept{league functions}
