% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-transactions.R
\name{league_transactions}
\alias{league_transactions}
\title{League transactions}
\usage{
league_transactions(lid, old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID.}

\item{old}{If \code{FALSE} (default), return only the current season's data. If
\code{TRUE}, return a list with all prior years.}

\item{...}{Additional API query arguments; use "week" as a shorthand for
"ScoringPeriodId" to display past rosters, etc.}
}
\value{
A tibble (or list) of league members.
}
\description{
Summary of transactions and roster changes made during a season.
}
\examples{
league_transactions(lid = 252353)
}
