% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tbl.R
\name{map_tbl}
\alias{map_tbl}
\title{Turn Nested Lists to Tibble Rows}
\usage{
map_tbl(l)
}
\arguments{
\item{l}{A list object whose sub-elements should be converted to rows and
bound into a single tibble.}
}
\value{
A nested list of API results.
}
\description{
Call \code{\link[purrr:map_df]{purrr::map_df()}} and \code{\link[tibble:as_tibble]{tibble::as_tibble()}} on a nested
list.
}
\examples{
m <- ffl_get(lid = 252353)
map_tbl(l = m$members)
}
