% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-scores.R
\name{match_scores}
\alias{match_scores}
\title{Fantasy matchup scores}
\usage{
match_scores(lid = getOption("lid"), old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID, defaults to \code{getOption("lid")} (see \code{\link[=set_lid]{set_lid()}}).}

\item{old}{If \code{FALSE} (default), return only the current season's data. If
\code{TRUE}, return a list with all prior seasons.}

\item{...}{Additional API query arguments; use "week" as a shorthand for
"ScoringPeriodId" to display past rosters, etc.}
}
\value{
A tibble (or list) of draft picks.
}
\description{
The weekly matchups between home and away teams, with score and winner.
}
\examples{
match_scores(lid = 252353)
}
