% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nfl_teams}
\alias{nfl_teams}
\title{2026 NFL Teams}
\format{
A data frame with 33 rows and 6 columns:
\describe{
\item{proTeamId}{Unique team ID}
\item{abbrev}{Professional team abbreviation}
\item{location}{Professional team geographic location}
\item{name}{Professional team full nickname}
\item{byeWeek}{Bye week, no game played}
\item{conference}{NFL conference}
...
}
}
\source{
\url{https://lm-api-reads.fantasy.espn.com/apis/v3/games/ffl/seasons/2026?view=proTeamSchedules_wl}
}
\usage{
nfl_teams
}
\description{
The 32 professional NFL teams as of the 2026 season.
}
\keyword{datasets}
