% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opponent-rank.R
\name{opponent_rank}
\alias{opponent_rank}
\title{Fantasy team roster}
\usage{
opponent_rank(lid = getOption("lid"), type = c("avg", "rank"))
}
\arguments{
\item{lid}{ESPN League ID, defaulted because all return the same data.}

\item{type}{Numeric data returned, either points average or rank.}
}
\value{
A tibble (or list) of players on a roster.
}
\description{
List all available players.
}
\examples{
opponent_rank()
}
