% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opponent-ranks.R
\name{opponent_ranks}
\alias{opponent_ranks}
\title{Fantasy team roster}
\usage{
opponent_ranks(lid = getOption("lid"))
}
\arguments{
\item{lid}{ESPN League ID, defaulted because all return the same data.}
}
\value{
A tibble (or list) of players on a roster.
}
\description{
List all available players.
}
\examples{
opponent_ranks()
}
