% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlook.R
\name{player_outlook}
\alias{player_outlook}
\title{Player outlooks}
\usage{
player_outlook(leagueId = ffl_id(), limit = 50)
}
\arguments{
\item{leagueId}{Numeric league ID or ESPN fantasy page URL. Defaults to
\code{getOption("fflr.leagueId")}. Function fails if no ID is found.}

\item{limit}{The limit of players to return. Use \code{""} or \code{NULL} to return
all. Defaults to 50, which is the default limit used by ESPN. Removing the
limit can make the request take a long time.}
}
\value{
A data frame of player outlooks by scoring period.
}
\description{
All available weekly ESPN outlook writeups for NFL players.
}
\examples{
player_outlook()
}
\seealso{
Other player functions:
\code{\link[=all_players]{all_players()}},
\code{\link[=list_players]{list_players()}},
\code{\link[=player_info]{player_info()}},
\code{\link[=player_lookup]{player_lookup()}},
\code{\link[=player_news]{player_news()}},
\code{\link[=recent_activity]{recent_activity()}},
\code{\link[=transaction_counter]{transaction_counter()}}
}
\concept{player functions}
