% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro-schedule.R
\name{pro_schedule}
\alias{pro_schedule}
\title{Professional schedule}
\usage{
pro_schedule(seasonId = ffl_year())
}
\arguments{
\item{seasonId}{Season schedule (2004-present), defaults to \code{\link[=ffl_year]{ffl_year()}}.}
}
\value{
Data frame of team opponents by week.
}
\description{
The opponents each team faces every week in a regular season.
}
\examples{
pro_schedule(seasonId = ffl_year(-2))
}
\seealso{
Other professional football functions:
\code{\link[=pro_events]{pro_events()}}
}
\concept{professional football functions}
