% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pro_teams}
\alias{pro_teams}
\title{The ESPN pro team IDs and their NFL abbreviations}
\format{A tibble with 32 rows and 2 variables:
\describe{
\item{team}{The ESPN team ID}
\item{pro}{The NFL team abbreviation}
...
}}
\usage{
pro_teams
}
\description{
The ESPN pro team IDs and their NFL abbreviations
}
\keyword{datasets}
