% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster_entry.R
\name{roster_entry}
\alias{roster_entry}
\title{Format a Roster Entry}
\usage{
roster_entry(entry)
}
\arguments{
\item{entry}{The \code{entry} index of \code{roster} list from a \code{team} list.}
}
\value{
A single row tibble roster entry.
}
\description{
Take a team's roster entry list and return a single row tibble.
}
\examples{
data <- ffl_get(lid = 252353, view = "roster", scoringPeriodId = 3)
entry <- data$teams[[5]]$roster$entries[[2]]
roster_entry(entry)
}
