% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-fill-fantasy.R
\name{scale_color_fantasy}
\alias{scale_color_fantasy}
\title{Extended qualitative color palettes}
\usage{
scale_color_fantasy()
}
\value{
A \code{ggplot2} object to define the fill of a plot.
}
\description{
This color palette combined the \code{Dark2} and \code{Set1} paletts from the Brewer
color system. Each of these palettes has 8 distinct qualitative colors. This
function combines the two palettes to provide enough distinct fill colors to
identify the 16 weeks of the NFL regular reason.
}
