% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_fantasy.R
\name{scale_fill_fantasy}
\alias{scale_fill_fantasy}
\title{Extended qualitative color palettes

This color palette combined the \code{Dark2} and \code{Set1} paletts from the Brewer
color system. Each of these palettes has 8 distinct qualitative colors. This
function combines the two palettes to provide enough distinct fill colors to
identify the 16 weeks of the NFL regular reason.}
\usage{
scale_fill_fantasy(variables)
}
\value{
A \code{ggplot2} object to define the fill of a plot.
}
\description{
Extended qualitative color palettes

This color palette combined the \code{Dark2} and \code{Set1} paletts from the Brewer
color system. Each of these palettes has 8 distinct qualitative colors. This
function combines the two palettes to provide enough distinct fill colors to
identify the 16 weeks of the NFL regular reason.
}
