% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{schedule_settings}
\alias{schedule_settings}
\title{League schedule settings}
\usage{
schedule_settings(leagueId = ffl_id(), leagueHistory = FALSE, ...)
}
\arguments{
\item{leagueId}{Numeric league ID or ESPN fantasy page URL. Defaults to
\code{getOption("fflr.leagueId")}. Function fails if no ID is found.}

\item{leagueHistory}{logical; Should the \code{leagueHistory} version of the API
be called? If \code{TRUE}, a list of results is returned, with one element for
each historical year of the league.}

\item{...}{Additional queries passed to \code{\link[httr:GET]{httr::GET()}}. Arguments are
converted to a named list and passed to \code{query} alongside \code{view}.}
}
\value{
A data frame of league schedule settings by season.
}
\description{
The length of a fantasy season and the match periods for each week.
}
\examples{
schedule_settings(leagueId = "42654852")
}
\seealso{
Other league settings functions:
\code{\link[=acquisition_settings]{acquisition_settings()}},
\code{\link[=draft_settings]{draft_settings()}},
\code{\link[=finance_settings]{finance_settings()}},
\code{\link[=league_info]{league_info()}},
\code{\link[=league_name]{league_name()}},
\code{\link[=league_size]{league_size()}},
\code{\link[=roster_settings]{roster_settings()}},
\code{\link[=scoring_settings]{scoring_settings()}},
\code{\link[=trade_settings]{trade_settings()}}
}
\concept{league settings functions}
