% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-lid.R
\name{set_lid}
\alias{set_lid}
\title{Set the ESPN league ID option}
\usage{
set_lid(lid = NULL, edit = FALSE, ...)
}
\arguments{
\item{lid}{ESPN league ID.}

\item{edit}{Should the \code{.Rprofile} be opened to edit? Ignored if the option
is already set. Call \code{\link[usethis:edit]{usethis::edit_r_profile()}} if installed.}

\item{...}{Arguments passed to \link[usethis:edit]{usethis::edit_r_profile}.}
}
\value{
Invisibly, the league ID.
}
\description{
Once the league ID option is set, the \code{lid} argument in most functions can be
ignored and will be automatically set. This function uses \code{\link[=options]{options()}} to
define the "lid" option. To persit this option, set \code{edit} to \code{TRUE} and
open the \code{.Rprofile} package, where \code{\link[=options]{options()}} can be added to define "lid".
}
\examples{
set_lid(252353)
}
