% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster.R
\name{start_roster}
\alias{start_roster}
\title{Starting roster}
\usage{
start_roster(roster)
}
\arguments{
\item{roster}{A roster data frame from \code{\link[=team_roster]{team_roster()}}.}
}
\value{
A data frame of starters on a roster.
}
\description{
The starting 9 man roster using standard roster slots. In the future this
function may be adapted to take roster slots from \code{\link[=roster_settings]{roster_settings()}}.
}
\examples{
rost <- team_roster(leagueId = "42654852")[[1]]
start_roster(rost)
}
\seealso{
Other roster functions:
\code{\link[=best_roster]{best_roster()}},
\code{\link[=evaluate_trade]{evaluate_trade()}},
\code{\link[=roster_score]{roster_score()}},
\code{\link[=team_roster]{team_roster()}},
\code{\link[=waiver_upgrades]{waiver_upgrades()}}
}
\concept{roster functions}
