% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchup.R
\name{tidy_schedule}
\alias{tidy_schedule}
\alias{tidy_matchups}
\title{Fantasy match schedule}
\usage{
tidy_schedule(leagueId = ffl_id(), leagueHistory = FALSE, ...)

tidy_matchups(...)
}
\arguments{
\item{leagueId}{Numeric league ID or ESPN fantasy page URL. Defaults to
\code{getOption("fflr.leagueId")}. Function fails if no ID is found.}

\item{leagueHistory}{logical; Should the \code{leagueHistory} version of the API
be called? If \code{TRUE}, a list of results is returned, with one element for
each historical year of the league.}

\item{...}{Additional queries passed to \code{\link[httr:GET]{httr::GET()}}. Arguments are
converted to a named list and passed to \code{query} alongside \code{view}.}
}
\value{
A data frame(s) of match opponents.
}
\description{
The opponents each team faces every week in a fantasy regular season.
Returned in a tidy format where each row is a single team with an indication
of home-away status. There are two rows per matchup, one for each team.
}
\examples{
tidy_schedule(leagueId = "42654852")
}
\seealso{
Other league functions:
\code{\link[=draft_recap]{draft_recap()}},
\code{\link[=league_members]{league_members()}},
\code{\link[=league_messages]{league_messages()}},
\code{\link[=league_standings]{league_standings()}},
\code{\link[=league_status]{league_status()}},
\code{\link[=league_teams]{league_teams()}},
\code{\link[=transaction_counter]{transaction_counter()}}
}
\concept{league functions}
