% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-settings.R
\name{waiver_settings}
\alias{waiver_settings}
\title{League waiver settings}
\usage{
waiver_settings(lid = getOption("lid"), old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID, defaults to \code{getOption("lid")} (see \code{\link[=set_lid]{set_lid()}}).}

\item{old}{If \code{FALSE} (default), return only the current season's data. If
\code{TRUE}, return a list with all prior seasons.}

\item{...}{Additional API query arguments; use "week" as a shorthand for
"ScoringPeriodId" to display past rosters, etc.}
}
\value{
A list or tibble of waiver settings.
}
\description{
League waiver settings
}
\examples{
waiver_settings(252353, old = TRUE)
}
