% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week-scores.R
\name{week_scores}
\alias{week_scores}
\title{Fantasy weekly scores}
\usage{
week_scores(lid = getOption("lid"), old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID, defaults to \code{getOption("lid")}. Use
\verb{options(lid = <League ID>)} to set your league ID; put that line in your
\code{.Rprofile} file to set the option at start up.}

\item{old}{If \code{FALSE} (default), return only the current season's data. If
\code{TRUE}, return a list with all prior seasons.}

\item{...}{Additional API query arguments. Use "week" and "year" as a
shorthand for "ScoringPeriodId" and "seasonId" respectively.}
}
\value{
A tibble of scores by week.
}
\description{
Weekly scores differ from [match_scores[]] in that matchups (especially
during the playoffs) often last multiple weeks.
}
\examples{
week_scores(lid = 252353)
}
