% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_matchup.R
\name{weekly_matchup}
\alias{weekly_matchup}
\title{Format a Weekly Matchup}
\usage{
weekly_matchup(matchup)
}
\arguments{
\item{entry}{An element of \code{schedule} list from a \code{mMatchup} view.}
}
\value{
A two-row tibble fora matchup's \code{HOME} and \code{AWAY} scores.
}
\description{
Take a weekly matchup and return tidy tibble of scores.
}
\examples{
data <- ffl_get(lid = 252353, view = "mMatchup")
matchup <- data$schedule[[5]]
weekly_matchup(matchup)
}
