% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{all_players}
\alias{all_players}
\title{All fantasy players}
\usage{
all_players(leagueId = ffl_id(), scoringPeriodId = ffl_week(), limit = 50)
}
\arguments{
\item{leagueId}{Numeric league ID or ESPN fantasy page URL. Defaults to
\code{getOption("fflr.leagueId")}. Function fails if no ID is found.}

\item{scoringPeriodId}{A scoring period to return, defaults to \code{\link[=ffl_week]{ffl_week()}}.}

\item{limit}{The limit of players to return. Use \code{""} or \code{NULL} to return
all. Defaults to 50, which is the default limit used by ESPN. Removing the
limit can make the request take a long time.}
}
\value{
A data frame of players.
}
\description{
List all available players.
}
\examples{
\dontrun{
all_players()
}
}
