% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-players.R
\name{all_players}
\alias{all_players}
\title{All fantasy players}
\usage{
all_players(lid = getOption("lid"), week = ffl_week())
}
\arguments{
\item{lid}{ESPN League ID, defaulted because all return the same data.}

\item{week}{The scoring period to return, defaults to \code{\link[=ffl_week]{ffl_week()}}.}
}
\value{
A tibble of players.
}
\description{
List all available players.
}
\examples{
all_players()
}
