% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.R
\name{best_roster}
\alias{best_roster}
\title{Sort the optimal fantasy roster}
\usage{
best_roster(
  leagueId = ffl_id(),
  scoringPeriodId = ffl_week(),
  useScore = c("actualScore", "projectedScore"),
  ...
)
}
\arguments{
\item{leagueId}{Numeric league ID or ESPN fantasy page URL. Defaults to
\code{getOption("fflr.leagueId")}. Function fails if no ID is found.}

\item{scoringPeriodId}{The scoring period for which rosters will be
optimized. If \code{scoringPeriodId} is the current week (the default), then
actual scoring might be incomplete (see \code{projectedScore} below).}

\item{useScore}{One of "projectedScore" or "actualScore" (default).}

\item{...}{Additional queries passed to \code{\link[httr:GET]{httr::GET()}}
(e.g., \code{scoringPeriodId}). Arguments are converted to a named list to be
passed alongside \code{view}.}
}
\value{
A dataframe (or list) with optimal rosters.
}
\description{
Uses the roster settings for each league to find the best possible
combinations of players to score the most fantasy points.
}
\examples{
best_roster(leagueId = "42654852", scoringPeriodId = 1)
}
