% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_roster.R
\name{best_roster}
\alias{best_roster}
\title{Find Optimal Roster}
\usage{
best_roster(roster)
}
\arguments{
\item{roster}{A roster tibble, as returned by \code{\link[=form_roster]{form_roster()}}.}
}
\value{
A tibble with the best roster, excluding bench.
}
\description{
Take a team's tibble (via \code{\link[=form_roster]{form_roster()}} and return the roster
lineup which would have scored the maximum number of points that week.
}
\examples{
data <- fantasy_roster(lid = 252353, scoringPeriodId = 9)
team <- data$teams[[5]]
roster <- form_roster(team)
best_roster(roster)
}
