% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_history}
\alias{combine_history}
\title{Combine league history with current season}
\usage{
combine_history(fun, ...)
}
\arguments{
\item{fun}{A function with the \code{leagueHistory} argument.}

\item{...}{Additional arguments passed to the function used in \code{fun}.}
}
\value{
A data frame of combined outputs.
}
\description{
Runs a function \code{fun} twice, once with the \code{leagueHistory} set to \code{TRUE} and
once set to \code{FALSE}. Combined the output of both runs into a single data
frame.
}
\examples{
combine_history(tidy_scores, leagueId = "252353")
}
