% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffl-api.R
\name{ffl_api}
\alias{ffl_api}
\title{Return list data from API JSON}
\usage{
ffl_api(lid = NULL, old = FALSE, view = NULL, ...)
}
\arguments{
\item{lid}{ESPN League ID.}

\item{old}{If \code{FALSE} (default), return only the current season's data. If
\code{TRUE}, return a list with all prior years.}

\item{view}{The API "view" for data to be returned.}

\item{...}{Additional API query arguments; use "week" as a shorthand for
"ScoringPeriodId" to display past rosters, etc.}
}
\description{
Return list data from API JSON
}
