% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{ffl_info}
\alias{ffl_info}
\title{Get fantasy football information}
\usage{
ffl_info()
}
\value{
A list of season information.
}
\description{
Information on the current fantasy football season.
}
\examples{
str(ffl_info())
}
