% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffl-info.R
\name{ffl_seasons}
\alias{ffl_seasons}
\title{List past fantasy seasons}
\usage{
ffl_seasons()
}
\value{
A tibble fantasy football seasons.
}
\description{
List past fantasy seasons
}
\examples{
ffl_seasons()
}
