% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{ffl_seasons}
\alias{ffl_seasons}
\title{List past fantasy football seasons}
\usage{
ffl_seasons()
}
\value{
A tibble of fantasy football seasons.
}
\description{
List past fantasy football seasons
}
\examples{
ffl_seasons()
}
