% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffl-info.R
\name{ffl_week}
\alias{ffl_week}
\title{Current fantasy scoring period}
\usage{
ffl_week(offset = 0)
}
\arguments{
\item{offset}{Add negative or positive values.}
}
\description{
Current fantasy scoring period
}
\examples{
Sys.time()
ffl_week(-1)
}
