% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fflr-package.R
\docType{package}
\name{fflr-package}
\alias{fflr}
\alias{fflr-package}
\title{fflr: A package for collecting ESPN fantasy football data}
\description{
The fflr package parses the JSON data returned by the ESPN v3 API into tidy
data frames for easy analysis. The package also includes data objects for the
NFL with players, teams, and the 2023 schedule. The league must first be made
viewable to the public by the league manager, which can be done on the basic
settings page on the ESPN website. Functions can then take the numeric league
ID found in the URL to return data from the API. This league ID can be set as
a global option named "fflr.leagueId" with \code{options()}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://k5cents.github.io/fflr/}
  \item \url{https://github.com/k5cents/fflr}
  \item Report bugs at \url{https://github.com/k5cents/fflr/issues}
}

}
\author{
\strong{Maintainer}: Kiernan Nicholls \email{k5cents@gmail.com} (\href{https://orcid.org/0000-0002-9229-7897}{ORCID}) [copyright holder]

}
\keyword{internal}
