% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_matchup.R
\name{form_matchup}
\alias{form_matchup}
\title{Format a Weekly Matchup}
\usage{
form_matchup(matchup)
}
\arguments{
\item{data}{The \code{schedule} element of the list object returned by
\code{\link[=fantasy_matchup]{fantasy_matchup()}}.}
}
\value{
A two-row tibble from matchup's \code{HOME} and \code{AWAY} scores.
}
\description{
Take a weekly matchup and return tidy tibble of scores.
}
\examples{
data <- fantasy_matchup(lid = 252353)
form_matchup(data$schedule[[10]])
}
