% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-settings.R
\name{league_info}
\alias{league_info}
\title{League information}
\usage{
league_info(lid, old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID.}

\item{old}{If \code{FALSE} (default), return only the current season's data. If
\code{TRUE}, return a list with all prior years.}

\item{...}{Additional API query arguments; use "week" as a shorthand for
"ScoringPeriodId" to display past rosters, etc.}
}
\value{
A list or tibble of season settings.
}
\description{
League information
}
\examples{
league_info(252353, old = TRUE)
}
