% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{league_info}
\alias{league_info}
\title{League information}
\usage{
league_info(leagueId = ffl_id(), leagueHistory = FALSE, ...)
}
\arguments{
\item{leagueId}{Numeric league ID or ESPN fantasy page URL. Defaults to
\code{getOption("fflr.leagueId")}. Function fails if no ID is found.}

\item{leagueHistory}{logical; Should the \code{leagueHistory} version of the API
be called? If \code{TRUE}, a list of results is returned, with one element for
each historical year of the league.}

\item{...}{Additional queries passed to \code{\link[httr:GET]{httr::GET()}}
(e.g., \code{scoringPeriodId}). Arguments are converted to a named list to be
passed alongside \code{view}.}
}
\value{
A data frame of league information by season.
}
\description{
Basic information on a ESPN fantasy football league, like the name, size,
and season length.
}
\examples{
league_info(leagueId = "42654852")
}
\seealso{
Other League settings: 
\code{\link{acquisition_settings}()},
\code{\link{draft_settings}()},
\code{\link{finance_settings}()},
\code{\link{league_name}()},
\code{\link{league_size}()},
\code{\link{roster_settings}()},
\code{\link{schedule_settings}()},
\code{\link{scoring_settings}()},
\code{\link{trade_settings}()}
}
\concept{League settings}
