% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-members.R
\name{league_members}
\alias{league_members}
\title{Fantasy league members}
\usage{
league_members(lid = getOption("lid"), old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID, defaults to \code{getOption("lid")} (see \code{\link[=set_lid]{set_lid()}}).}

\item{old}{If \code{FALSE} (default), return only the current season's data. If
\code{TRUE}, return a list with all prior seasons.}

\item{...}{Additional API query arguments; use "week" as a shorthand for
"ScoringPeriodId" to display past rosters, etc.}
}
\value{
A tibble (or list) of league members.
}
\description{
Return the ESPN user names, unique owner IDs, and their manager status.
}
\examples{
league_members(lid = 252353)
}
