% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-members.R
\name{league_members}
\alias{league_members}
\title{Fantasy league members}
\usage{
league_members(lid, old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID.}

\item{old}{If \code{FALSE} (default), return only the current season's draft
history. If \code{TRUE}, return a list with all prior drafts.}

\item{...}{Additional API query arguments passed to \code{\link[httr:GET]{httr::GET()}}.}
}
\value{
A tibble (or list) of league members.
}
\description{
Return the ESPN user names, unique owner IDs, and their manager status.
}
\examples{
league_members(lid = 252353)
}
