% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-settings.R
\name{league_name}
\alias{league_name}
\title{League name}
\usage{
league_name(lid, old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID.}

\item{old}{If \code{FALSE} (default), return only the current season's data. If
\code{TRUE}, return a list with all prior years.}

\item{...}{Additional API query arguments; use "week" as a shorthand for
"ScoringPeriodId" to display past rosters, etc.}
}
\value{
A character vector.
}
\description{
League name
}
\examples{
league_name(252353)
}
