% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{league_name}
\alias{league_name}
\title{League name}
\usage{
league_name(leagueId = ffl_id(), leagueHistory = FALSE, ...)
}
\arguments{
\item{leagueId}{Numeric league ID or ESPN fantasy page URL. Defaults to
\code{getOption("fflr.leagueId")}. Function fails if no ID is found.}

\item{leagueHistory}{logical; Should the \code{leagueHistory} version of the API
be called? If \code{TRUE}, a list of results is returned, with one element for
each historical year of the league.}

\item{...}{Additional queries passed to \code{\link[httr:GET]{httr::GET()}}. Arguments are
converted to a named list and passed to \code{query} alongside \code{view}.}
}
\value{
A character vector.
}
\description{
League name
}
\examples{
league_name(leagueId = "42654852")
}
\seealso{
Other league settings functions: 
\code{\link{acquisition_settings}()},
\code{\link{draft_settings}()},
\code{\link{finance_settings}()},
\code{\link{league_info}()},
\code{\link{league_size}()},
\code{\link{roster_settings}()},
\code{\link{schedule_settings}()},
\code{\link{scoring_settings}()},
\code{\link{trade_settings}()}
}
\concept{league settings functions}
