% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-settings.R
\name{league_size}
\alias{league_size}
\title{League team size}
\usage{
league_size(lid, old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID.}

\item{old}{If \code{FALSE} (default), return only the current season's data. If
\code{TRUE}, return a list with all prior years.}

\item{...}{Additional API query arguments; use "week" as a shorthand for
"ScoringPeriodId" to display past rosters, etc.}
}
\value{
A named integer vector.
}
\description{
League team size
}
\examples{
league_size(252353, old = TRUE)
}
