% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{league_size}
\alias{league_size}
\title{League size}
\usage{
league_size(leagueId = ffl_id(), leagueHistory = FALSE, ...)
}
\arguments{
\item{leagueId}{Numeric league ID or ESPN fantasy page URL. Defaults to
\code{getOption("fflr.leagueId")}. Function fails if no ID is found.}

\item{leagueHistory}{logical; Should the \code{leagueHistory} version of the API
be called? If \code{TRUE}, a list of results is returned, with one element for
each historical year of the league.}

\item{...}{Arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}} for parsing.}
}
\value{
A data frame of league size by season.
}
\description{
League size
}
\examples{
league_size(leagueId = "42654852")
league_size(leagueId = "252353", leagueHistory = TRUE)
}
\seealso{
Other League settings: 
\code{\link{draft_settings}()},
\code{\link{fee_settings}()},
\code{\link{league_info}()},
\code{\link{league_name}()},
\code{\link{roster_settings}()},
\code{\link{schedule_settings}()},
\code{\link{score_settings}()},
\code{\link{trade_settings}()},
\code{\link{waiver_settings}()}
}
\concept{League settings}
