% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{league_status}
\alias{league_status}
\title{League status}
\usage{
league_status(leagueId = ffl_id(), leagueHistory = FALSE, ...)
}
\arguments{
\item{leagueId}{Numeric league ID or ESPN fantasy page URL. Defaults to
\code{getOption("fflr.leagueId")}. Function fails if no ID is found.}

\item{leagueHistory}{logical; Should the \code{leagueHistory} version of the API
be called? If \code{TRUE}, a list of results is returned, with one element for
each historical year of the league.}

\item{...}{Additional queries passed to \code{\link[httr:GET]{httr::GET()}}
(e.g., \code{scoringPeriodId}). Arguments are converted to a named list to be
passed alongside \code{view}.}
}
\value{
A data frame of league status by season.
}
\description{
Current information about a league: the date activated, current week,
starting week, final week, past seasons, teams joined, and waiver status.
}
\examples{
league_status(leagueId = "42654852")
}
