% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-teams.R
\name{league_teams}
\alias{league_teams}
\title{Fantasy league teams}
\usage{
league_teams(lid, old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID.}

\item{old}{If \code{FALSE} (default), return only the current season's draft
history. If \code{TRUE}, return a list with all prior drafts.}

\item{...}{Additional API query arguments passed to \code{\link[httr:GET]{httr::GET()}}.}
}
\value{
A tibble (or list) of league teams.
}
\description{
Return the fantasy team names and their owners.
}
\examples{
league_teams(lid = 252353)
}
