% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{league_teams}
\alias{league_teams}
\title{Bus Schedule at Stop}
\format{
A tibble with 1 row per bus departure and 8 variables:
\describe{
\item{abbrev}{Team name abbreviation}
\item{id}{League team ID}
\item{location}{Team location}
\item{nickname}{Team nickname}
\item{owners}{List of all team owner user IDs}
}
}
\usage{
league_teams(seasonId = 2021, leagueId = ffl_id(), leagueHistory = FALSE)
}
\arguments{
\item{seasonId}{Integer year of NFL season. By default, the season is
currently set to 2021. Use a recent year or set \code{leagueHistory} to \code{TRUE}
to obtain past data.}

\item{leagueId}{Unique league ID, defaults to \code{\link[=ffl_id]{ffl_id()}}.}

\item{leagueHistory}{logical; Should the \code{leagueHistory} version of the API
be called? If \code{TRUE}, a list of results is returned, with one element for
each historical year of the league.}
}
\value{
Data frame containing scheduled arrival information.
}
\description{
Returns a set of buses scheduled at a stop for a given date.
}
\examples{
league_teams(leagueId = "42654852")
}
\concept{League information}
