% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-teams.R
\name{league_teams}
\alias{league_teams}
\title{Fantasy league teams}
\usage{
league_teams(lid = getOption("lid"), old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID, defaults to \code{getOption("lid")}. Use
\verb{options(lid = <League ID>)} to set your league ID; put that line in your
\code{.Rprofile} file to set the option at start up.}

\item{old}{If \code{FALSE} (default), return only the current season's data. If
\code{TRUE}, return a list with all prior seasons.}

\item{...}{Additional API query arguments. Use "week" and "year" as a
shorthand for "ScoringPeriodId" and "seasonId" respectively.}
}
\value{
A tibble (or list) of league teams.
}
\description{
The teams in a league and their owners.
}
\examples{
league_teams(lid = 252353)
}
