% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live-scoring.R
\name{live_scoring}
\alias{live_scoring}
\title{Live matchup scoreboard}
\usage{
live_scoring(lid = getOption("lid"), old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID, defaults to \code{getOption("lid")} (see \code{\link[=set_lid]{set_lid()}}).}

\item{old}{If \code{FALSE} (default), return only the current season's data. If
\code{TRUE}, return a list with all prior seasons.}

\item{...}{Additional API query arguments; use "week" as a shorthand for
"ScoringPeriodId" to display past rosters, etc.}
}
\value{
A tibble of scores.
}
\description{
Live matchup scoreboard
}
\examples{
live_scoring(252353)
}
