% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live-scoring.R
\name{live_scoring}
\alias{live_scoring}
\title{Live matchup scoreboard}
\usage{
live_scoring(lid = getOption("lid"), yet = FALSE)
}
\arguments{
\item{lid}{ESPN League ID, defaults to \code{getOption("lid")}. Use
\verb{options(lid = <League ID>)} to set your league ID; put that line in your
\code{.Rprofile} file to set the option at start up.}

\item{yet}{If \code{TRUE}, \code{\link[=pro_schedule]{pro_schedule()}} and the "mRoster" view are called to
determine how many starting players have \emph{yet} started playing.}
}
\value{
A tibble of scores.
}
\description{
The current and projected score for each ongoing match.
}
\examples{
live_scoring(252353)
}
