% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-schedule.R
\name{match_schedule}
\alias{match_schedule}
\title{Match schedule}
\usage{
match_schedule(lid = getOption("lid"), old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID, defaults to \code{getOption("lid")} (see \code{\link[=set_lid]{set_lid()}}).}

\item{old}{If \code{FALSE} (default), return only the current season's data. If
\code{TRUE}, return a list with all prior seasons.}

\item{...}{Additional API query arguments; use "week" as a shorthand for
"ScoringPeriodId" to display past rosters, etc.}
}
\value{
A tibble (or list) of match opponents.
}
\description{
Match schedule
}
\examples{
match_schedule(lid = 252353)
}
