% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{moves_summary}
\alias{moves_summary}
\title{League transactions}
\usage{
moves_summary(leagueId = ffl_id(), leagueHistory = FALSE, ...)
}
\arguments{
\item{leagueId}{Numeric league ID or ESPN fantasy page URL. Defaults to
\code{getOption("fflr.leagueId")}. Function fails if no ID is found.}

\item{leagueHistory}{logical; Should the \code{leagueHistory} version of the API
be called? If \code{TRUE}, a list of results is returned, with one element for
each historical year of the league.}

\item{...}{Arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}} for parsing.}
}
\value{
A data frame of transaction counts by team.
}
\description{
Summary of transactions and roster changes made during a season by team.
}
\examples{
moves_summary(leagueId = "42654852")
}
