% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-transactions.R
\name{moves_summary}
\alias{moves_summary}
\title{League transactions}
\usage{
moves_summary(lid = getOption("lid"), old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID, defaults to \code{getOption("lid")} (see \code{\link[=set_lid]{set_lid()}}).}

\item{old}{If \code{FALSE} (default), return only the current season's data. If
\code{TRUE}, return a list with all prior seasons.}

\item{...}{Additional API query arguments. Use "week" and "year" as a
shorthand for "ScoringPeriodId" and "seasonId" respectively.}
}
\value{
A tibble (or list) of league members.
}
\description{
Summary of transactions and roster changes made during a season.
}
\examples{
moves_summary(lid = 252353)
}
