% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nfl_players}
\alias{nfl_players}
\title{2020 NFL Players}
\format{
A data frame with 1,059 rows and 5 variables:
\describe{
\item{id}{Unique ESPN player ID}
\item{first}{First name}
\item{last}{Last name}
\item{pro}{Professional NFL team}
\item{pos}{Position: QB, RB, WR, TE, D/ST}
...
}
}
\source{
\url{https://fantasy.espn.com/apis/v3/games/ffl/seasons/2020/segments/0/leagues/252353?view=kona_player_info}
}
\usage{
nfl_players
}
\description{
All available ESPN fantasy football players as of the 2020 season, week 4.
}
\keyword{datasets}
