% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nfl_players}
\alias{nfl_players}
\title{2020 NFL Players}
\format{
A data frame with 1,059 rows and 5 variables:
\describe{
\item{id}{Unique ESPN player ID}
\item{first}{First name}
\item{last}{Last name}
\item{pro}{Professional NFL team}
\item{pos}{Position: QB, RB, WR, TE, D/ST}
\item{jersey}{Jersey number}
\item{weight}{Weight in even pounds}
\item{height}{Weight in even inches}
\item{age}{Current age in year}
\item{birth}{Date of birth}
\item{debut}{Season debuted in league}
...
}
}
\source{
\url{http://sports.core.api.espn.com/v2/sports/football/leagues/nfl/seasons/2020/athletes/}
}
\usage{
nfl_players
}
\description{
All available ESPN fantasy football players as of the 2020 season, week 4.
}
\keyword{datasets}
