% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nfl_schedule}
\alias{nfl_schedule}
\title{2020 NFL Schedule}
\format{
A data frame with 1,059 rows and 5 variables:
\describe{
\item{year}{Season year}
\item{week}{Matchup period}
\item{pro}{Team abbreviation}
\item{opp}{Team opponent}
\item{home}{Team the home?}
\item{kickoff}{Match date and time}
...
}
}
\source{
\url{https://fantasy.espn.com/apis/v3/games/ffl/seasons/2020/segments/0/leagues/252353?view=kona_player_info}
}
\usage{
nfl_schedule
}
\description{
All available ESPN fantasy football players as of the 2020 season, week 4.
}
\keyword{datasets}
