% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nfl_schedule}
\alias{nfl_schedule}
\title{2020 NFL Schedule}
\format{
A data frame with 1,059 rows and 5 variables:
\describe{
\item{year}{Season year}
\item{week}{Matchup period}
\item{pro}{Team abbreviation}
\item{opp}{Team opponent}
\item{home}{Team the home?}
\item{kickoff}{Match date and time}
...
}
}
\source{
\url{https://fantasy.espn.com/apis/v3/games/ffl/seasons/2020?view=proTeamSchedules_wl}
}
\usage{
nfl_schedule
}
\description{
The 2020 NFL season schedule by team, as of October 5th. Due to COVID-19 game
delays, the schedule is more in flux than usual. See \code{\link[=pro_schedule]{pro_schedule()}} for the
most current version.
}
\keyword{datasets}
