% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nfl_schedule}
\alias{nfl_schedule}
\title{2021 NFL Schedule}
\format{
A data frame with 544 rows and 6 variables:
\describe{
\item{seasonId}{Season year}
\item{scoringPeriodId}{Matchup period}
\item{proTeam}{Team abbreviation}
\item{opponent}{Team opponent}
\item{isHome}{Team the home?}
\item{date}{Match date and time}
...
}
}
\source{
\url{https://fantasy.espn.com/apis/v3/games/ffl/seasons/2021?view=proTeamSchedules_wl}
}
\usage{
nfl_schedule
}
\description{
The 2021 NFL season schedule by team, as of September 10th.
}
\keyword{datasets}
