% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nfl_teams}
\alias{nfl_teams}
\title{Team abbreviations in the NFL}
\format{
A data frame with 1,056 rows and 5 variables:
\describe{
\item{team}{Unique yearly team ID}
\item{abbrev}{Team abbreviation}
...
}
}
\usage{
nfl_teams
}
\description{
Team abbreviations in the NFL
}
\keyword{datasets}
