% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nfl_teams}
\alias{nfl_teams}
\title{2020 NFL Teams}
\format{
A data frame with 1,056 rows and 5 variables:
\describe{
\item{team}{Unique team ID}
\item{abbrev}{Team abbreviation}
\item{location}{Team geographic location}
\item{name}{Team full nickname}
\item{bye}{Bye week, no game played}
\item{conf}{NFL conference}
...
}
}
\source{
\url{https://fantasy.espn.com/apis/v3/games/ffl/seasons/2020?view=proTeamSchedules_wl}
}
\usage{
nfl_teams
}
\description{
The 32 professional NFL teams as of the 2020 season.
}
\keyword{datasets}
