% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opponent-ranks.R
\name{opponent_ranks}
\alias{opponent_ranks}
\title{NFL team performance against positions}
\usage{
opponent_ranks(lid = getOption("lid"))
}
\arguments{
\item{lid}{ESPN League ID, defaulted because all return the same data.}
}
\value{
A tibble of performance stats.
}
\description{
The average opposition team point differential by position.
}
\examples{
opponent_ranks()
}
