% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opponents.R
\name{opponent_ranks}
\alias{opponent_ranks}
\title{NFL team performance against positions}
\usage{
opponent_ranks(leagueId = ffl_id())
}
\arguments{
\item{leagueId}{Numeric league ID or ESPN fantasy page URL. Defaults to
\code{getOption("fflr.leagueId")}. Function fails if no ID is found.}
}
\value{
A data frame of team performance against position.
}
\description{
The average opposition team point differential by position.
}
\examples{
opponent_ranks()
}
