% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire.R
\name{player_acquire}
\alias{player_acquire}
\title{Roster acquisition history}
\usage{
player_acquire(leagueId = ffl_id(), leagueHistory = FALSE, ...)
}
\arguments{
\item{leagueId}{Numeric league ID or ESPN fantasy page URL. Defaults to
\code{getOption("fflr.leagueId")}. Function fails if no ID is found.}

\item{leagueHistory}{logical; Should the \code{leagueHistory} version of the API
be called? If \code{TRUE}, a list of results is returned, with one element for
each historical year of the league.}

\item{...}{Additional queries passed to \code{\link[httr:GET]{httr::GET()}}
(e.g., \code{scoringPeriodId}). Arguments are converted to a named list to be
passed alongside \code{view}.}
}
\value{
A data frame of roster players with acquisition method and date.
}
\description{
The date and method of each player's acquisition onto a fantasy roster.
}
\examples{
player_acquire(leagueId = "42654852")
}
