% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-info.R
\name{player_info}
\alias{player_info}
\title{Individual player information}
\usage{
player_info(id, row = FALSE)
}
\arguments{
\item{id}{A single player ID number.}

\item{row}{Should list be converted to a single row data frame?}
}
\value{
A tibble of players.
}
\description{
Individual player information
}
\examples{
str(player_info(15847))
}
