% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{player_info}
\alias{player_info}
\title{Individual player information}
\usage{
player_info(playerId)
}
\arguments{
\item{playerId}{A single player ID number.}
}
\value{
A list or row of a single player's information.
}
\description{
Individual player information
}
\examples{
player_info(playerId = 15847)
}
