% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-news.R
\name{player_news}
\alias{player_news}
\title{Player news}
\usage{
player_news(id, parse = FALSE)
}
\arguments{
\item{id}{A vector of player ID numbers.}

\item{parse}{Should HTML stories be parsed with \code{\link[xml2:read_xml]{xml2::read_html()}}?}
}
\value{
A tibble of news stories.
}
\description{
The free and premium ESPN stories on given players. A maximum of 50 stories
can be returned at a time.
}
\examples{
player_news(15847)
}
