% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-outlook.R
\name{player_outlook}
\alias{player_outlook}
\title{Player outlooks}
\usage{
player_outlook(lid = getOption("lid"))
}
\arguments{
\item{lid}{ESPN League ID, defaulted because all return the same data.}
}
\value{
A tibble of player outlooks.
}
\description{
All available weekly ESPN outlook writeups for NFL players.
}
\examples{
player_outlook()
}
