% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro-events.R
\name{pro_events}
\alias{pro_events}
\title{Professional games}
\usage{
pro_events()
}
\value{
A data frame of NFL events.
}
\description{
Data on the status of NFL games, including scores and odds, kickoff time,
and broadcast information.
}
\examples{
pro_events()
}
\seealso{
Other professional football functions: 
\code{\link{pro_schedule}()}
}
\concept{professional football functions}
