% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro-schedule.R
\name{pro_schedule}
\alias{pro_schedule}
\title{Professional NFL schedule}
\usage{
pro_schedule(year = ffl_year())
}
\arguments{
\item{year}{Season schedule (2004-present), defaults to \code{\link[=ffl_year]{ffl_year()}}.}
}
\value{
Tibble of NFL team opponents by week.
}
\description{
The opponents each team faces every week in an NFL regular season.
}
\examples{
pro_schedule(year = ffl_year(-2))
}
