% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster-entry.R
\name{roster_entry}
\alias{roster_entry}
\title{Format a Roster Entry}
\usage{
roster_entry(entry)
}
\arguments{
\item{entry}{The \code{entry} element of \code{roster} list from a \code{team} list.}
}
\value{
A single row tibble roster entry.
}
\description{
Take a team's roster entry list and return a single row tibble.
This function is used in \code{\link[=form_roster]{form_roster()}} to combine the entries from an
entire team into a single tibble.
}
\examples{
data <- fantasy_roster(lid = 252353, scoringPeriodId = 3)
roster_entry(entry = data$teams[[5]]$roster$entries[[2]])
}
