% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best-roster.R
\name{roster_score}
\alias{roster_score}
\title{Sum Roster Score}
\usage{
roster_score(roster)
}
\arguments{
\item{roster}{A roster tibble, as returned by \code{\link[=form_roster]{form_roster()}} or
\code{\link[=best_roster]{best_roster()}}.}
}
\value{
A double.
}
\description{
For a given roster tibble, sum the starting scores.
}
\examples{
data <- fantasy_roster(lid = 252353, scoringPeriodId = 2)
start_roster <- form_roster(data$teams[[5]])
best_roster <- best_roster(start_roster)
start_score <- roster_score(start_roster)
best_score <- roster_score(best_roster)
start_score/best_score
}
