% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best-roster.R
\name{roster_score}
\alias{roster_score}
\title{Sum a roster score}
\usage{
roster_score(roster)
}
\arguments{
\item{roster}{A roster tibble from \code{\link[=team_roster]{team_roster()}} or \code{\link[=best_roster]{best_roster()}}.}
}
\value{
A score as double.
}
\description{
For a given roster tibble, sum the starting scores.
}
\examples{
# asses performance compared to optimal
start <- team_roster(lid = 252353, week = 1)[[5]]
best <- best_roster(start)
roster_score(start)/roster_score(best)
}
