% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best-roster.R
\name{start_roster}
\alias{start_roster}
\title{Starting roster}
\usage{
start_roster(roster)
}
\arguments{
\item{roster}{A roster tibble from \code{\link[=team_roster]{team_roster()}}.}
}
\value{
A tibble of starters on a roster.
}
\description{
The starting 9 man roster using standard roster slots. In the future this
function may be adapted to take roster slots from \code{\link[=roster_settings]{roster_settings()}}.
}
\examples{
start_roster(team_roster(lid = 252353, week = 1)[[7]])
}
