% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-teams.R
\name{team_abbrev}
\alias{team_abbrev}
\title{Convert team ID to abbreviation}
\usage{
team_abbrev(id, lid = getOption("lid"), ...)
}
\arguments{
\item{id}{A integer vector of team IDs to convert.}

\item{lid}{ESPN League ID passed to \code{\link[=league_teams]{league_teams()}} or the table itself.}

\item{...}{Arguments passed to \code{\link[=league_teams]{league_teams()}}.}
}
\value{
A character vector of team abbreviations.
}
\description{
Convert team ID to abbreviation
}
\examples{
team_abbrev(252353, id = 5)
}
