% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-teams.R
\name{team_abbrev}
\alias{team_abbrev}
\title{Convert team ID to abbreviation}
\usage{
team_abbrev(id, teams = league_teams(lid = getOption("lid")))
}
\arguments{
\item{id}{A integer vector of team numbers to convert.}

\item{teams}{A table of teams from \code{\link[=league_teams]{league_teams()}}.}
}
\value{
A factor vector of team abbreviations.
}
\description{
Convert team ID to abbreviation
}
\examples{
team_abbrev(id = 6, teams = league_teams(lid = 252353))
}
