% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrev.R
\name{team_abbrev}
\alias{team_abbrev}
\title{Convert team ID to abbreviation}
\usage{
team_abbrev(teamId, teams = league_teams(leagueId = ffl_id()))
}
\arguments{
\item{teamId}{A integer vector of team numbers to convert.}

\item{teams}{A table of teams, like that from \code{\link[=league_teams]{league_teams()}}.}
}
\value{
A factor vector of team abbreviations.
}
\description{
Convert team ID to abbreviation
}
\examples{
team_abbrev(teamId = 2, teams = league_teams(leagueId = "42654852"))
}
