% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team-roster.R
\name{team_roster}
\alias{team_roster}
\title{Fantasy team roster}
\usage{
team_roster(lid, old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID.}

\item{old}{If \code{FALSE} (default), return only the current season's draft
history. If \code{TRUE}, return a list with all prior drafts.}

\item{...}{Additional API query arguments passed to \code{\link[httr:GET]{httr::GET()}}.}
}
\value{
A tibble (or list) of players on a roster.
}
\description{
The players on a team roster with position, injury status, and playing stats.
}
\examples{
team_roster(lid = 252353)
}
