% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_roster.R
\name{team_roster}
\alias{team_roster}
\title{Format a Team Roster}
\usage{
team_roster(team)
}
\arguments{
\item{team}{The \code{team} index of \code{\link[=ffl_get]{ffl_get()}} response.}
}
\value{
A tibble with a row for every roster entry.
}
\description{
Take a team's list roster entries and applies \code{\link[=roster_entry]{roster_entry()}}
to return a single tibble per team.
}
\examples{
data <- ffl_get(lid = 252353, view = "roster", scoringPeriodId = 3)
team <- data$teams[[5]]
team_roster(team)
}
