% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team-roster.R
\name{team_roster}
\alias{team_roster}
\title{Fantasy team roster}
\usage{
team_roster(lid = getOption("lid"), old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID, defaults to \code{getOption("lid")}. Use
\verb{options(lid = <League ID>)} to set your league ID; put that line in your
\code{.Rprofile} file to set the option at start up.}

\item{old}{If \code{FALSE} (default), return only the current season's data. If
\code{TRUE}, return a list with all prior seasons.}

\item{...}{Additional API query arguments. Use "week" and "year" as a
shorthand for "ScoringPeriodId" and "seasonId" respectively.}
}
\value{
A tibble (or list) of players on a roster.
}
\description{
The players on a team roster with position, injury status, and playing stats.
}
\examples{
team_roster(lid = 252353)
}
