% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly-matchups.R
\name{weekly_matchups}
\alias{weekly_matchups}
\title{Fantasy matchups}
\usage{
weekly_matchups(lid, old = FALSE, ...)
}
\arguments{
\item{lid}{ESPN League ID.}

\item{old}{If \code{FALSE} (default), return only the current season's draft
history. If \code{TRUE}, return a list with all prior drafts.}

\item{...}{Additional API query arguments passed to \code{\link[httr:GET]{httr::GET()}}.}
}
\value{
A tibble (or list) of draft picks.
}
\description{
The weekly matchups between home and away teams, with score and winner.
}
\examples{
weekly_matchups(lid = 252353)
}
